Move requests and responses to presentation layer
