+++ src/main/java/com/ite/itea/application/dto/ChairsDto.java
@@ -1,6 +1,6 @@
 package com.ite.itea.application.dto;
 
-import com.ite.itea.domain.request.ItemNameRequest;
+import com.ite.itea.presentation.request.ItemNameRequest;
 
 public class ChairsDto extends ItemDto {
