+++ src/main/java/com/ite/itea/presentation/CheckoutController.java
@@ -6,10 +6,10 @@ import com.ite.itea.application.dto.PicturesDto;
 import com.ite.itea.application.dto.TablesDto;
 import com.ite.itea.domain.CheckoutCalculator;
 import com.ite.itea.application.dto.*;
-import com.ite.itea.domain.request.ItemNameRequest;
-import com.ite.itea.domain.request.ItemRequest;
-import com.ite.itea.domain.request.OrderRequest;
-import com.ite.itea.domain.response.ReceiptResponse;
+import com.ite.itea.presentation.request.ItemNameRequest;
+import com.ite.itea.presentation.request.ItemRequest;
+import com.ite.itea.presentation.request.OrderRequest;
+import com.ite.itea.presentation.response.ReceiptResponse;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
