+++ src/main/java/com/ite/itea/presentation/request/ItemNameRequest.java
@@ -1,4 +1,4 @@
-package com.ite.itea.domain.request;
+package com.ite.itea.presentation.request;
 
 public enum ItemNameRequest {
     PictureFinland, PictureNorway, PictureSweden, ChairElsa, ChairLars, ChairKnut, TableLotta, TableLola
