+++ src/main/java/com/ite/itea/presentation/request/ItemRequest.java
@@ -1,4 +1,4 @@
-package com.ite.itea.domain.request;
+package com.ite.itea.presentation.request;
 
 public record ItemRequest(ItemNameRequest name, int amount, long price) {
