+++ src/main/java/com/ite/itea/presentation/response/ReceiptResponse.java
@@ -1,4 +1,4 @@
-package com.ite.itea.domain.response;
+package com.ite.itea.presentation.response;
 
 public record ReceiptResponse(long priceInCents, String text) {
